requirements.txt
flask
flask_sqlalchemy
flask_cors
gunicorn
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0

flask
flask_sqlalchemy
flask_cors
gunicorn
psycopg2-binary
python-dotenv

